+++ HISTORY.rst
@@ -2,7 +2,15 @@
 History
 =======
 
-1.0.0 (2025-03-05)
+1.0.1 (2025-03-09)
+------------------
+* Read discovery date from 'name' in instead of from 'packed_designation'. Packed designation losses his date meaning when asteroid get numbered 'name' field has discovery date meaning while it is provisional (no given name).
+* tqdm to show progress
+* fix json date serialization
+* format float,integer and hex acording specification
+
+
+1.0.0 (2025-03-07)
 ------------------
 * Change internal. All objects now are store with its type (not strings) thus to_numeric() is not used ever
 * Put convinience fn outside of main class